src/publish_frame/requirements.txt
kafka-python
pydantic
python-dotenv
watchdog